Implemented some 64bit safety changes from TyrQuake

+++ source/d_modech.c
@@ -42,8 +42,8 @@ void D_Patch (void)
 
 	if (!protectset8)
 	{
-		Sys_MakeCodeWriteable ((int)D_PolysetAff8Start,
-						     (int)D_PolysetAff8End - (int)D_PolysetAff8Start);
+		Sys_MakeCodeWriteable ((unsigned long)D_PolysetAff8Start,
+						     (unsigned long)D_PolysetAff8End - (unsigned long)D_PolysetAff8Start);
 		protectset8 = true;
 	}
